airbyte-integrations/connectors/source-microsoft-sharepoint-user/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.5.1"
name = "source-microsoft-sharepoint-user"
description = "Source implementation for Microsoft SharePoint."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/microsoft-sharepoint"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
[[tool.poetry.packages]]
include = "source_microsoft_sharepoint_user"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
Office365-REST-Python-Client ="==2.5.12"
airbyte-cdk = {version="^4", extras=["file-based",]}

[tool.poetry.scripts]
source-microsoft-sharepoint-user = "source_microsoft_sharepoint_user.run:run"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.6.1"
requests-mock = "^1.11.0"
pytest = "^6.1"
